Refactor: Remove "custom" namespace from CLI imports
Relocated files and updated imports to remove the "custom" namespace, aligning all modules under a unified "cli" structure. This enhances the maintainability and consistency of the project structure, reducing unnecessary directory nesting.
