+++ src/terry/presentation/cli/messages/dir_activate_message.py
