+++ src/terry/presentation/cli/messages/files_select_message.py
