+++ src/terry/presentation/cli/messages/move_resizing_rule.py
