+++ src/terry/presentation/cli/messages/path_delete_message.py
