+++ src/terry/presentation/cli/messages/tf_command_output.py
