+++ src/terry/presentation/cli/screens/add_file/main.py
@@ -9,7 +9,7 @@ from textual.screen import ModalScreen
 from textual.widgets import Label, Input
 
 from terry.infrastructure.file_system.exceptions import CreateFileException, CreateDirException
-from terry.presentation.cli.custom.widgets.modal_control_label import ModalControlLabel
+from terry.presentation.cli.widgets.modal_control_label import ModalControlLabel
 
 
 class Control(ModalControlLabel):
