+++ src/terry/presentation/cli/screens/tf_format/main.py
@@ -6,8 +6,8 @@ from textual.screen import ModalScreen
 from textual.widgets import Label, RadioSet, RadioButton
 
 from terry.domain.terraform.core.entities import FormatScope
-from terry.presentation.cli.custom.messages.tf_format_action_request import FormatActionRequest
-from terry.presentation.cli.custom.widgets.modal_control_label import (
+from terry.presentation.cli.messages.tf_format_action_request import FormatActionRequest
+from terry.presentation.cli.widgets.modal_control_label import (
     ModalControlLabel,
 )
