+++ src/terry/presentation/cli/screens/tf_validate/main.py
@@ -6,10 +6,10 @@ from textual.widgets import Static, Rule
 
 from terry.domain.terraform.core.entities import TerraformValidateSettingsAttributes, ValidateSettings
 from terry.presentation.cli.commands_descriptions import VALIDATE_DESCRIPTION, VALIDATE_NO_TESTS_DESCRIPTION
-from terry.presentation.cli.custom.messages.tf_validate_action_request import ValidateActionRequest
-from terry.presentation.cli.custom.widgets.buttons.open_file_navigator_modal_button import FileNavigatorModalButton
-from terry.presentation.cli.custom.widgets.form.checkbox_settings_block import CheckboxSettingBlock
-from terry.presentation.cli.custom.widgets.modal_control_label import ModalControlLabel
+from terry.presentation.cli.messages.tf_validate_action_request import ValidateActionRequest
+from terry.presentation.cli.widgets.buttons.open_file_navigator_modal_button import FileNavigatorModalButton
+from terry.presentation.cli.widgets.form.checkbox_settings_block import CheckboxSettingBlock
+from terry.presentation.cli.widgets.modal_control_label import ModalControlLabel
 from terry.presentation.cli.screens.base.base_tf_settings_screen import BaseTfSettingsModalScreen
 from terry.presentation.cli.screens.file_system_navigation.main import FileSystemSelectionValidationRule
