+++ src/terry/presentation/cli/widgets/buttons/delete_button.py
