+++ src/terry/presentation/cli/widgets/buttons/open_file_navigator_modal_button.py
