+++ src/terry/presentation/cli/widgets/buttons/sidebar_button.py
