+++ src/terry/presentation/cli/widgets/buttons/view_secret_field_button.py
