+++ src/terry/presentation/cli/widgets/clickable_icon.py
