+++ src/terry/presentation/cli/widgets/clickable_tf_action_label.py
