+++ src/terry/presentation/cli/widgets/form/collapsible_info_settings_block.py
@@ -1,34 +1,24 @@
 from textual import on
 from textual.app import ComposeResult
-from textual.containers import Container, Horizontal
+from textual.containers import Horizontal, Container
 from textual.css.query import NoMatches
-from textual.widgets import Input, Static
+from textual.widgets import Label, Static
 
-from terry.presentation.cli.custom.widgets.clickable_icon import ClickableIcon
-from terry.presentation.cli.custom.widgets.form.collapsible_with_no_title import CollapsibleWithNoTitle
+from terry.presentation.cli.widgets.clickable_icon import ClickableIcon
+from terry.presentation.cli.widgets.form.collapsible_with_no_title import CollapsibleWithNoTitle
 
 
-class TextInputBlock(Container):
+class CollapsibleInfoBlock(Container):
     DEFAULT_CSS = """
-    TextInputBlock {
+    CollapsibleInfoBlock {
         height: auto;
 
         & > Horizontal {
-            width: 100%;
-            height: auto;
-            padding_bottom: 1;
-            padding_top: 0;
+        height: auto;
 
-            & > Input {
-                width: 65%
-            }
-            & > ClickableIcon {
-                overflow-x: hidden;
-                width: 35%;
-                height: 3;
-                padding: 1 1;
-            }
-
+        & > Label {
+            padding-right: 1;
+        }
       }
 
       & > CollapsibleWithNoTitle {
@@ -55,14 +45,12 @@ class TextInputBlock(Container):
         self.setting_name = setting_name
         self.label = label
         self.description = description
-        self._input = None
         super().__init__(**kwargs)
 
     def compose(self) -> ComposeResult:
-        self._input = Input()
         with Horizontal(id=f"{self.setting_name}_block"):
-            yield ClickableIcon(f"{self.label} (i):", name=self.setting_name)
-            yield self._input
+            yield Label(self.label)
+            yield ClickableIcon("(i)", name=self.setting_name)
         with CollapsibleWithNoTitle(collapsed=True, title="", id=f"{self.setting_name}_toggle"):
             yield Static(self.description)
 
@@ -83,9 +71,3 @@ class TextInputBlock(Container):
         except NoMatches:
             return
         collapsible.collapsed = not collapsible.collapsed
-
-    @property
-    def content(self):
-        if self._input is None:
-            return None
-        return self._input.value
