+++ src/terry/presentation/cli/widgets/form/collapsible_with_no_title.py
