+++ tests/integration/screens/main/test_content.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 from textual.widgets import TextArea
 
-from terry.presentation.cli.custom.messages.files_select_message import FileSelect
+from terry.presentation.cli.messages.files_select_message import FileSelect
 from terry.settings import ANIMATION_SPEED
 from tests.integration.utils import click
