+++ tests/integration/screens/main/test_main.py
@@ -4,7 +4,7 @@ import pytest
 from watchdog.events import FileSystemEvent, EVENT_TYPE_MODIFIED, EVENT_TYPE_CREATED
 
 from terry.infrastructure.terraform.core.exceptions import TerraformFormatException
-from terry.presentation.cli.custom.messages.tf_format_action_request import FormatActionRequest
+from terry.presentation.cli.messages.tf_format_action_request import FormatActionRequest
 from terry.presentation.cli.screens.tf_format.main import FormatScope
 from terry.presentation.cli.screens.main.containers.content import Content
 from terry.settings import DEFAULT_THEME, CommandStatus
